requirements.txt
# Database
psycopg2-binary>=2.9.9

# Video/Audio Download
yt-dlp>=2023.12.30

# AI APIs
anthropic>=0.18.0
openai-whisper>=20231117

# Data Analysis
scikit-learn>=1.3.2
pandas>=2.1.4
numpy>=1.26.2

# Visualization
matplotlib>=3.8.2
seaborn>=0.13.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
